refactor for light sds
